order, taxes and some security issues

+++ branches/nyholmsolutions/plugins/sfsBasketPlugin/modules/basket/lib/BaseBasketActions.class.php
@@ -299,7 +299,7 @@ class BaseBasketActions extends sfActions
             $basket = $this->getUser()->getBasket();
             $basketProduct = BasketProductPeer::retrieveByPk($request->getParameter('id'));
             
-            if ($basketProduct !== null) {
+            if ($basketProduct !== null && $basketProduct->getBasketId() == $basket->getId()) {
                 $basketProduct->delete();
             }
         }
